Move file observer into utils package
